fix(deno): move types to better named file
